tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.interpolate.zerosize.fillText.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.gradient.interpolate.zerosize.fillText</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.gradient.interpolate.zerosize.fillText</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);

  var g = ctx.createLinearGradient(50, 25, 50, 25); // zero-length line (undefined direction)
  g.addColorStop(0, '#f00');
  g.addColorStop(1, '#f00');
  ctx.fillStyle = g;
  ctx.font = '100px sans-serif';
  ctx.fillText("AA", 0, 50);
  _assertGreen(ctx, 100, 50);
  t.done();

});
</script>
